+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Change log
   is ignored by OneDrive & reports a notice. Omit this option.
 - `\Krizalys\Onedrive\Proxy\DriveItemProxy::upload()`: The `'contentType'`
   option is ignored by OneDrive & reports a notice. Omit this option.
+- Updated Krizalys\Onedrive\Proxy\ImageProxy.
 
 [2.5.1] - 2019-10-06
 --------------------